auth-service/requirements.txt
fastapi
uvicorn
motor
passlib[bcrypt]
python-jose[cryptography]
email-validator
pytest
pytest-asyncio
